New data
Here is a Machine Learning example that I created using the California House data set. In this Machine Learning I use RMSE, MAE, and MAPE as evaluation metrics because the data set used belongs to the regression category.

After testing, it turned out that the best model is using XGBoost because it has a MAPE value of 18.2%, so it can be categorized as 'good forecasting' (Lewis, 1982).